PCI: rockchip: rk3399: Ensure that PERST is asserted for at least 100 ms
The PCI Express Card Electromechanical Specification specifies:
"On power up, the deassertion of PERST# is delayed 100 ms (TPVPERL) from
the power rails achieving specified operating limits."

Add a sleep of 100 ms before deasserting PERST, in order to ensure that
we are compliant with the spec.

Also wait for a "Minimum PERST# inactive to PCI Express link out of
electrical idle" time.

Signed-off-by: Soeren Moch <smoch@web.de>